Rename lib -> src
