+++ package.json
@@ -2,7 +2,7 @@
   "name": "hyperscript-html",
   "version": "1.0.0",
   "description": "A fast, simple, hyperscript implementation for generating HTML.",
-  "main": "lib/index",
+  "main": "src/index",
   "repository": "https://github.com/arccoza/hyperscript-html",
   "bugs": "https://github.com/arccoza/hyperscript-html/issues",
   "author": "Adrien de Pierres <adrien@arccoza.com>",
